requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
websockets==12.0
docker==7.0.0
psutil==5.9.6
transformers==4.35.2
torch==2.1.1
protobuf==4.25.1 